feat: refactor code for esm.mjs file
